// shared/types/enums/activityLevel.ts
export enum ActivityLevel {
    Sedentary = 0, // Little or no exercise
    Light = 1, // Light exercise/sports 1-3 days/week
    Moderately = 2, // Moderate exercise/sports 3-5 days/week
    High = 3, // Hard exercise/sports 6-7 days a week
    Extreme = 4, // Very hard exercise/sports & a physical job or training
}